+++ test/dyno_node.py
@@ -2,16 +2,49 @@
 import redis
 from node import Node
 from redis_node import RedisNode
-from dyno_cluster import DynoCluster
+
+from plumbum import BG
+from plumbum import local
+
+dynomite_bin = local.get('./test/_binaries/dynomite', 'src/dynomite')
 
 class DynoNode(Node):
-    def __init__(self, ip, port=8102,
-                 dnode_port=8101, data_store_port=1212):
-        super(DynoNode, self).__init__(ip, port)
+    #TODO: Decouple from Redis and make it more generic
+    def __init__(self, spec, seeds_list):
+        super(DynoNode, self).__init__(spec.ip, spec.port)
         self.name="Dyno" + self.name
         self.conf_file = None
-        self.dnode_port = dnode_port
-        self.data_store_node = RedisNode(ip, data_store_port)
+        self.spec = spec
+        self.seeds_list = seeds_list
+        self.dnode_port = spec.dnode_port
+        self.data_store_node = RedisNode(self.ip, spec.data_store_port)
+        self.logfile = 'logs/dynomite_{}.log'.format(self.ip)
+        self.proc_future = None
+
+    def launch(self):
+        # Write the configuration to a file according to 'self.spec'.
+        config_filename = self.spec.write_config(self.seeds_list)
+
+        # Launch the underlying data store process first.
+        self.data_store_node.launch()
+
+        # Launch the dynomite process
+        self.proc_future = dynomite_bin['-o', self.logfile, '-c', config_filename,
+            '-v6'] & BG(-9)
+
+    def teardown(self):
+        # First teardown the Dynomite node.
+        self.proc_future.proc.kill()
+        self.proc_future.wait()
+
+        # Lastly, teardown the storage node.
+        self.data_store_node.teardown()
+
+    def __enter__(self):
+        self.launch()
+
+    def __exit__(self, type_, value, traceback):
+        self.teardown()
 
     def get_connection(self):
         # should return the connection to the dyno port not the redis
